+++ src/api/admin/Swagger/views/scripts.js
@@ -1,1 +1,0 @@
-// SwaggerAdmin scripts